Update patch set 2
Patch Set 2:

One last comment, wsgi service expects the api_paste_config to be in api-paste.ini. However we are shiping it in /etc/novajoin/join-api-paste.ini. Given that we are moving the config to a specific directory, i'd change file name to api-paste.ini so that it works with default configuration and we don't need to change the parameter.

Patch-set: 2