Switch unit tests to stestr
Novajoin switched to stestr in [1]. After that, it moved some
functional tests in [2]. We need to use stestr, so it will detect
the right tests and not try to run functional tests.

Note this cannot be cherry-picked from https://review.rdoproject.org/r/17274
due to the differences between rpm-master and rocky-rdo.

[1] - https://review.openstack.org/573154
[2] - https://review.openstack.org/615817

Change-Id: I567e866c84eca999d405621cd6491e366c705fe9